move new funcs into more appropriate file
