+++ pkg/ezkube/resource_id.go
@@ -1,10 +1,20 @@
 package ezkube
 
 import (
+	"strings"
+	"sync"
+
+	"github.com/rotisserie/eris"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const ClusterAnnotation = "cluster.solo.io/cluster"
+
+var builderPool = sync.Pool{
+	New: func() any {
+		return &strings.Builder{}
+	},
+}
 
 // ResourceId represents a global identifier for a k8s resource.
 type ResourceId interface {
@@ -17,6 +27,20 @@ type ClusterResourceId interface {
 	GetName() string
 	GetNamespace() string
 	GetAnnotations() map[string]string
+}
+
+// internal struct that satisfies ResourceId interface
+type resourceId struct {
+	name      string
+	namespace string
+}
+
+func (id resourceId) GetName() string {
+	return id.name
+}
+
+func (id resourceId) GetNamespace() string {
+	return id.namespace
 }
 
 // internal struct needed to create helper func that converts ref to struct that satisfies ClusterResourceId interface
@@ -83,4 +107,59 @@ func SetClusterName(obj client.Object, cluster string) {
 		obj.SetAnnotations(map[string]string{})
 	}
 	obj.GetAnnotations()[ClusterAnnotation] = cluster
+}
+
+// KeyWithSeparator constructs a string consisting of the field values of the given resource id,
+// separated by the given separator. It can be used to construct a unique key for a resource.
+func KeyWithSeparator(id ResourceId, separator string) string {
+	b := builderPool.Get().(*strings.Builder)
+	defer func() {
+		b.Reset()
+		builderPool.Put(b)
+	}()
+	// When kubernetes objects are passed in here, a call to the GetX() functions will panic, so
+	// this will return "<unknown>" always if the input is nil.
+	if id == nil {
+		return "<unknown>"
+	}
+	b.WriteString(id.GetName())
+	b.WriteString(separator)
+	b.WriteString(id.GetNamespace())
+	b.WriteString(separator)
+	// handle the possibility that clusterName could be set either as an annotation (new way)
+	// or as a field (old way pre-k8s 1.25)
+	if clusterId, ok := id.(ClusterResourceId); ok {
+		clusterNameByAnnotation := GetClusterName(clusterId)
+		if clusterNameByAnnotation != "" {
+			b.WriteString(clusterNameByAnnotation)
+			return b.String()
+		}
+	}
+	if deprecatedClusterId, ok := id.(interface{ GetClusterName() string }); ok {
+		b.WriteString(deprecatedClusterId.GetClusterName())
+	}
+	return b.String()
+}
+
+// ResourceIdFromKeyWithSeparator converts a key back into a ResourceId, using the given separator.
+// Returns an error if it cannot be converted.
+func ResourceIdFromKeyWithSeparator(key string, separator string) (ResourceId, error) {
+	parts := strings.Split(key, separator)
+	if len(parts) == 2 {
+		return resourceId{
+			name:      parts[0],
+			namespace: parts[1],
+		}, nil
+
+	} else if len(parts) == 3 {
+		return clusterResourceId{
+			name:      parts[0],
+			namespace: parts[1],
+			annotations: map[string]string{
+				ClusterAnnotation: parts[2],
+			},
+		}, nil
+	} else {
+		return nil, eris.Errorf("could not convert key %s with separator %s into resource id; unexpected number of parts: %d", key, separator, len(parts))
+	}
 }
